+++ data/duke.txt
@@ -1,2 +1,1 @@
-todo asdas 
-deadline asdsjgghalsf  /by 2/2/2020 1800
+event run  /from 10pm  /to 11pm
